Create tutor
This file is an attempt to create a tool to assist my learning.

The goal is simply to read a line of text from an input file. Print that line to the terminal, Read user input (typing a string to match the read line, repeat on failure and continue to the next line as available.